requirements.txt
nanoid
pyopenssl
secrets
coincurve
sha3
Pillow
pyjwt
cryptography